move actions map
